chore: refactor ModuleStatusCache to its own file
