+++ lib/get-config.js
@@ -364,103 +364,6 @@ function _determineConfigForSeverity(config) {
   }
 }
 
-class ModuleStatusCache {
-  constructor(config, configPath) {
-    this.config = config;
-    this.configPath = configPath || '';
-    this.cache = {
-      pending: {},
-      ignore: {},
-    };
-    this.processCWD = process.cwd();
-  }
-
-  lookupPending(moduleId) {
-    if (!moduleId || !this.config.pending) {
-      return false;
-    }
-    if (!this.cache.pendingLookup) {
-      this.cache.pendingLookup = this._extractPendingCache();
-    }
-    if (!(moduleId in this.cache.pending)) {
-      const fullPathModuleId = path.resolve(this.processCWD, moduleId);
-      this.cache.pending[moduleId] = this.cache.pendingLookup[fullPathModuleId];
-    }
-    return this.cache.pending[moduleId];
-  }
-
-  lookupIgnore(moduleId) {
-    if (!(moduleId in this.cache.ignore)) {
-      const ignores = this.config['ignore'] || [];
-      this.cache.ignore[moduleId] = ignores.find((match) => match(moduleId));
-    }
-    return Boolean(this.cache.ignore[moduleId]);
-  }
-
-  _extractPendingCache() {
-    const list = this.config.pending;
-    const byFullModuleId = {};
-
-    if (!list) {
-      return byFullModuleId;
-    }
-
-    for (const item of list) {
-      if (typeof item === 'string') {
-        const fullPath = this.resolveFullModuleId(item);
-        byFullModuleId[fullPath] = true;
-      } else if (item.moduleId) {
-        const fullPath = this.resolveFullModuleId(item.moduleId);
-        byFullModuleId[fullPath] = item;
-      }
-    }
-
-    return byFullModuleId;
-  }
-
-  resolveFullModuleId(moduleId) {
-    if (!this._baseDirBasedOnConfigPath) {
-      this._baseDirBasedOnConfigPath = path.resolve(this.processCWD, path.dirname(this.configPath));
-    }
-    return path.resolve(this._baseDirBasedOnConfigPath, moduleId);
-  }
-}
-
-let configModuleCacheMap = new WeakMap();
-
-/**
- * Returns the config in conjunction with overrides configuration.
- * @param {*} config
- * @param {*} filePath
- */
-function getConfigForFile(config, options) {
-  if (!configModuleCacheMap.has(config)) {
-    configModuleCacheMap.set(config, new ModuleStatusCache(config, options.configPath));
-  }
-  let moduleStatusCache = configModuleCacheMap.get(config);
-  let filePath = options.filePath;
-  let configuredRules = config.rules;
-  let overrides = config.overrides;
-
-  let fileConfig = Object.assign({}, config, {
-    pendingStatus: moduleStatusCache.lookupPending(options.moduleId),
-    shouldIgnore: moduleStatusCache.lookupIgnore(options.moduleId),
-  });
-
-  if (filePath && overrides) {
-    let overridesRules = {};
-    overrides.forEach((override) => {
-      const isFileMatch = override.files && micromatch.isMatch(filePath, override.files);
-      if (isFileMatch) {
-        overridesRules = Object.assign(overridesRules, override.rules);
-      }
-    });
-    // If there is an override present, then the overridden ruleset takes precedence.
-    fileConfig.rules = Object.assign({}, configuredRules, overridesRules);
-  }
-  return fileConfig;
-}
-
 function determineRuleConfig(ruleData) {
   let ruleConfig = {
     severity: ruleData === false ? IGNORE_SEVERITY : ERROR_SEVERITY,
@@ -542,7 +445,6 @@ module.exports = {
   ERROR_SEVERITY,
   IGNORE_SEVERITY,
   WARNING_SEVERITY,
-  getConfigForFile,
   getProjectConfig,
   getRuleFromString,
   resolveProjectConfig,
