+++ lib/linter.js
@@ -1,8 +1,8 @@
 const { parse, transform } = require('ember-template-recast');
 
+const ModuleStatusCache = require('./-private/module-status-cache');
 const {
   getProjectConfig,
-  getConfigForFile,
   getRuleFromString,
   WARNING_SEVERITY,
   ERROR_SEVERITY,
@@ -56,6 +56,7 @@ class Linter {
 
       this.config.rules[name] = config;
     }
+    this._moduleStatusCache = new ModuleStatusCache(this.config, this.options.configPath);
   }
 
   /**
@@ -233,7 +234,7 @@ class Linter {
     }
 
     let currentSource = options.source;
-    let fileConfig = getConfigForFile(this.config, options);
+    let fileConfig = this._moduleStatusCache.getConfigForFile(options);
 
     let pending = fileConfig.pending;
     let ruleNames = new Set(fixableIssues.map((r) => r.rule));
@@ -266,7 +267,7 @@ class Linter {
    */
   verify(options) {
     let results = [];
-    let fileConfig = getConfigForFile(this.config, options);
+    let fileConfig = this._moduleStatusCache.getConfigForFile(options);
     let pendingStatus = fileConfig.pendingStatus;
 
     if (fileConfig.shouldIgnore) {
